# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Baby Health Application API
  version: 1.0.0
  description: Multi-baby health record API (India). Reference values are returned only with provenance and only when their release gate is CLEARED.
    The API never returns a computed milk target for healthy infants.
  contact:
    name: Product owner
servers:
- url: https://api.example.in
  description: Production (India region) - placeholder
security:
- bearerAuth: []
...
paths:
...
  /v1/babies/{babyId}/prescriptions:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    get:
      summary: List prescriptions
      tags:
      - Prescriptions
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Prescription'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    post:
      summary: Create prescriptions record
      tags:
      - Prescriptions
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionCreate'
  /v1/babies/{babyId}/prescriptions/{id}:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - $ref: '#/components/parameters/Id'
    get:
      summary: Get prescriptions record
      tags:
      - Prescriptions
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    patch:
      summary: Update prescriptions record
      tags:
      - Prescriptions
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
        '422':
          $ref: '#/components/responses/RuleViolation'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionPatch'
    delete:
      summary: Soft-delete prescriptions record
      tags:
      - Prescriptions
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
  /v1/babies/{babyId}/medicines:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    get:
      summary: List medicines
      tags:
      - Medicines
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Medicine'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    post:
      summary: Create medicines record
      tags:
      - Medicines
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medicine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineCreate'
  /v1/babies/{babyId}/medicines/{id}:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - $ref: '#/components/parameters/Id'
    get:
      summary: Get medicines record
      tags:
      - Medicines
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medicine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    patch:
      summary: Update medicines record
      tags:
      - Medicines
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medicine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
        '422':
          $ref: '#/components/responses/RuleViolation'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicinePatch'
    delete:
      summary: Soft-delete medicines record
      tags:
      - Medicines
      x-authorization: OWNER, GUARDIAN
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '412':
          $ref: '#/components/responses/VersionMismatch'
  /v1/babies/{babyId}/medicines/{id}/schedule:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - $ref: '#/components/parameters/Id'
    put:
      summary: Set medicine reminder schedule
      tags:
      - Medicines
      x-authorization: OWNER, GUARDIAN
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicineSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineSchedule'
  /v1/babies/{babyId}/medicines/{id}/doses:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - $ref: '#/components/parameters/Id'
    get:
      summary: List doses
      tags:
      - Medicines
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/From'
      - $ref: '#/components/parameters/To'
      - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MedicineDose'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    post:
      summary: Record a dose as given/skipped
      tags:
      - Medicines
      x-authorization: OWNER, GUARDIAN, CAREGIVER
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicineDose'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineDoseCreate'
...
  /v1/babies/{babyId}/allergies/confirm-none:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    post:
      summary: Confirm no known allergies
      tags:
      - Allergies
      x-authorization: OWNER, GUARDIAN
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  nka_confirmed_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token <= 15 min; authorization resolved per baby via membership + database row-level security.
  parameters:
    BabyId:
      name: babyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resources of babies without membership return 404.
...
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    From:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      schema:
        type: string
        format: date-time
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum:
        - occurred_at
        - -occurred_at
        - created_at
        - -created_at
        default: -occurred_at
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  schemas:
...
    Prescription:
      allOf:
      - $ref: '#/components/schemas/PrescriptionCreate'
      - $ref: '#/components/schemas/Audit'
      - type: object
        properties:
          id:
            type: string
            format: uuid
...